✨ query 외부파일 분리
